prune on demande

+++ resources/start-ifbcat.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 set -e
 
+PRUNE=0
+if [[ "$1" == "--prune" ]]; then
+    PRUNE=1
+fi
+
 cd /var/ifbcat-src
 
 export CI_COMMIT_SHA=$(git rev-parse HEAD)
@@ -10,7 +15,9 @@ echo "Starting IFB Catalogue with commit ${CI_COMMIT_SHA}"
 
 docker compose down --remove-orphans
 
+if [[ $PRUNE -eq 1 ]]; then
 docker image prune -a --force --filter "until=672h"
+fi
 
 docker compose pull --quiet --ignore-pull-failures
